--- dataset/XRayDatasetAll.py
@@ -1,116 +1,0 @@
-import os
-import json
-
-# external library
-import cv2
-import numpy as np
-from sklearn.model_selection import GroupKFold
-
-# torch
-import torch
-from torch.utils.data import Dataset
-
-
-IMAGE_ROOT = "./data/train/DCM"
-LABEL_ROOT = "./data/train/outputs_json"
-
-CLASSES = [
-    'finger-1', 'finger-2', 'finger-3', 'finger-4', 'finger-5',
-    'finger-6', 'finger-7', 'finger-8', 'finger-9', 'finger-10',
-    'finger-11', 'finger-12', 'finger-13', 'finger-14', 'finger-15',
-    'finger-16', 'finger-17', 'finger-18', 'finger-19', 'Trapezium',
-    'Trapezoid', 'Capitate', 'Hamate', 'Scaphoid', 'Lunate',
-    'Triquetrum', 'Pisiform', 'Radius', 'Ulna',
-]
-
-CLASS2IND = {v: i for i, v in enumerate(CLASSES)}
-
-IND2CLASS = {v: k for k, v in CLASS2IND.items()}
-
-pngs = {
-    os.path.relpath(os.path.join(root, fname), start=IMAGE_ROOT)
-    for root, _dirs, files in os.walk(IMAGE_ROOT)
-    for fname in files
-    if os.path.splitext(fname)[1].lower() == ".png"
-}
-
-jsons = {
-    os.path.relpath(os.path.join(root, fname), start=LABEL_ROOT)
-    for root, _dirs, files in os.walk(LABEL_ROOT)
-    for fname in files
-    if os.path.splitext(fname)[1].lower() == ".json"
-}
-
-jsons_fn_prefix = {os.path.splitext(fname)[0] for fname in jsons}
-pngs_fn_prefix = {os.path.splitext(fname)[0] for fname in pngs}
-
-assert len(jsons_fn_prefix - pngs_fn_prefix) == 0
-assert len(pngs_fn_prefix - jsons_fn_prefix) == 0
-
-pngs = sorted(pngs)
-jsons = sorted(jsons)
-
-
-class XRayDatasetAll(Dataset):
-    def __init__(self, is_train=True, transforms=None, fold=0):
-        _filenames = np.array(pngs)
-        _labelnames = np.array(jsons)
-
-        ############# Train 800 ###############
-        if is_train:
-            # 모든 train 데이터
-            filenames = list(_filenames)
-            labelnames = list(_labelnames)
-        else:
-            # validation 데이터
-            filenames = []
-            labelnames = []
-
-        self.filenames = filenames
-        self.labelnames = labelnames
-        self.is_train = is_train
-        self.transforms = transforms
-
-    def __len__(self):
-        return len(self.filenames)
-
-    def __getitem__(self, item):
-        image_name = self.filenames[item]
-        image_path = os.path.join(IMAGE_ROOT, image_name)
-
-        image = cv2.imread(image_path)
-        image = image / 255.
-
-        label_name = self.labelnames[item]
-        label_path = os.path.join(LABEL_ROOT, label_name)
-
-        label_shape = tuple(image.shape[:2]) + (len(CLASSES), )
-        label = np.zeros(label_shape, dtype=np.uint8)
-
-        with open(label_path, "r") as f:
-            annotations = json.load(f)
-        annotations = annotations["annotations"]
-
-        for ann in annotations:
-            c = ann["label"]
-            class_ind = CLASS2IND[c]
-            points = np.array(ann["points"])
-            class_label = np.zeros(image.shape[:2], dtype=np.uint8)
-            cv2.fillPoly(class_label, [points], 1)
-            label[..., class_ind] = class_label
-
-        if self.transforms is not None:
-            inputs = {"image": image, "mask": label} if self.is_train else {"image": image}
-            result = self.transforms(**inputs)
-
-            image = result["image"]
-            label = result["mask"] if self.is_train else label
-
-        # to tenser will be done later
-        image = image.transpose(2, 0, 1)
-        label = label.transpose(2, 0, 1)
-
-        image = torch.from_numpy(image).float()
-        label = torch.from_numpy(label).float()
-
-        return image, label
